fix: Resolve circular import issues from package/file naming conflicts
Fixed critical circular import issues caused by directory/file naming
conflicts in the AUTOSAR model structure.

## Problem
Python's import system cannot distinguish between a package (directory with
__init__.py) and a module (.py file) with the same name. When both exist,
Python prefers the package, causing import confusion and circular dependency
errors.

## Changes
- Removed duplicate AutosarTopLevelStructure.py (kept directory package)
- Added ModeInBswModuleDescriptionInstanceRef class to BswOverview.py
- Removed empty directories that shadowed .py files:
  - ECUCParameterDefTemplate/
  - SystemTemplate/InstanceRefs/
  - Composition/InstanceRefs/

## Test Coverage
✅ All 2263 unit tests pass
✅ No flake8 E9,F63,F7,F82 errors
✅ Clean build with pip install . succeeds
✅ No logic changes - only structural cleanup

Closes #347
