--- src/armodel/models/M2/AUTOSARTemplates/BswModuleTemplate/BswOverview/ModeInBswModuleDescriptionInstanceRef.py
@@ -1,17 +1,0 @@
-"""
-This module defines mode in BSW module description instance reference in AUTOSAR.
-"""
-
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
-
-
-class ModeInBswModuleDescriptionInstanceRef(RefType):
-    """
-    Represents a reference to a mode in BSW module description.
-    """
-
-    def __init__(self):
-        """
-        Initializes the ModeInBswModuleDescriptionInstanceRef.
-        """
-        super().__init__()
